# Fix timing for GPU execution