### Before 
`Avg request latency: 138887 usec (overhead 42 usec + queue 80 usec + compute input 13897 usec + compute infer 10288 usec + compute output 114580 usec)`

### After
`Avg request latency: 128321 usec (overhead 36 usec + queue 75 usec + compute input 20 usec + compute infer 128107 usec + compute output 83 usec)`