@tanmayv25 2021-05-06
> While I did confirm that the forward pass should be async and the assumption for it to be synchronized before reading output tensors is valid.

So, you confirm the reading first output synchronizes stream? 

> Did you test this on a multi-gpu device and see how it impacts perf?

I haven't yet.. I can try that if you believe this can be an issue..
I am using resnet50_fp32_libtorch model from our perf_model_store.