@CoderHam 2021-05-06
Btw cudaLaunchHostFunc was added in cuda 10, should there be a related check around this?