@CoderHam 2021-05-06
> So, you confirm the reading first output synchronizes stream?

No, my reading did not confirm that the first o/p synchronizes the stream. I was only able to confirm the forward call being async.