@deadeyegoodwin 2021-05-06
> 
> 
> I guess that makes sense. I was thinking of a case where a user might want to build Triton + pytorch_backend manually for cuda < 10.

We don't need to directly support that and can leave it to that user to sort out.