@tanmayv25 2021-05-12
@deadeyegoodwin  I did some digging and think my latest change should behave as expected or close enough.

The pytorch JIT uses our provided device(CPU/GPU) from [`map_location`](http://man.hubwiz.com/docset/PyTorch.docset/Contents/Resources/Documents/jit.html#torch.jit.load). In case it is CPU the operations are performed synchronously but when the location is GPU the operations are enqueued on a cuda stream. One can use CudaGuard to relay operations through specific streams. In our backend we don't use one hence, it uses a default cuda stream.
@CoderHam  Correct if my understanding appears erroneous.

There is a hiccup in cudaLaunchHostFunc but I think we should be okay.
> Using CUDA 11.3, and 466.11 Driver.
> When Hardware-accelerated GPU scheduling is OFF, cudaLaunchHostFunc serializes (blocks work on all streams).
> When Hardware-accelerated GPU scheduling is ON, cudaLaunchHostFunc does not block work on other streams (correct behavior).

Source: https://forums.developer.nvidia.com/t/does-cudalaunchhostfunc-block-work-added-to-all-streams/170259/19