@CoderHam 2021-05-12
> @CoderHam Correct if my understanding appears erroneous.
Based on my reading I believe your understanding is correct. If CudaGuard is the preferred way in PyTorch to relay instructions we can consider adding the same. However as I understand the cuda stream used in CopyBuffer is independent from the default stream being used by PyTorch. By using CudaGuard will we be able to (and want to) use a single stream for both?